+++ openapi/data_source_factory.go
@@ -35,8 +35,8 @@ func (d dataSourceFactory) createTerraformDataSource() (*schema.Resource, error)
 		return nil, err
 	}
 	return &schema.Resource{
-		Schema: s,
-		Read:   d.read,
+		Schema:      s,
+		ReadContext: crudWithContext(d.read, schema.TimeoutRead, d.openAPIResource.GetResourceName()),
 	}, nil
 }
