+++ openapi/data_source_instance_factory.go
@@ -29,8 +29,8 @@ func (d dataSourceInstanceFactory) createTerraformInstanceDataSource() (*schema.
 		return nil, err
 	}
 	return &schema.Resource{
-		Schema: s,
-		Read:   d.read,
+		Schema:      s,
+		ReadContext: crudWithContext(d.read, schema.TimeoutRead, d.openAPIResource.GetResourceName()),
 	}, nil
 }
