+++ openapi/resource_factory_test.go
@@ -105,76 +105,6 @@ func TestCreateTerraformResource(t *testing.T) {
 	})
 }
 
-func TestCRUDWithContext(t *testing.T) {
-	Convey("Given a resource factory configured with a create timeout", t, func() {
-		createTimeout := 1 * time.Second
-		r := resourceFactory{
-			openAPIResource: &specStubResource{
-				name: "cdn_v1",
-			},
-		}
-		Convey("When crudWithContext is called with the create function which returns successfully", func() {
-			stubCreateFunction := func(data *schema.ResourceData, i interface{}) error {
-				return nil // this means the function returned successfully
-			}
-			contextAwareFunc := r.crudWithContext(stubCreateFunction, schema.TimeoutCreate)
-			Convey("Then the returned function which is context aware should not timeout and return an empty diagnosis", func() {
-				ctx := context.Background()
-				ctx, cancel := context.WithTimeout(ctx, createTimeout)
-				defer cancel()
-				diagnosis := contextAwareFunc(ctx, &schema.ResourceData{}, nil)
-				So(diagnosis, ShouldBeEmpty)
-			})
-		})
-	})
-	Convey("Given a resource factory configured with a create timeout", t, func() {
-		createTimeout := 1 * time.Second
-		r := resourceFactory{
-			openAPIResource: &specStubResource{
-				name: "cdn_v1",
-			},
-		}
-		Convey("When crudWithContext is called with the create function which returns an error", func() {
-			expectedError := "some error"
-			stubCreateFunction := func(data *schema.ResourceData, i interface{}) error {
-				return errors.New(expectedError)
-			}
-			contextAwareFunc := r.crudWithContext(stubCreateFunction, schema.TimeoutCreate)
-			Convey("Then the returned function which is context aware should not timeout and return the error from the create function", func() {
-				ctx := context.Background()
-				ctx, cancel := context.WithTimeout(ctx, createTimeout)
-				defer cancel()
-				diagnosis := contextAwareFunc(ctx, &schema.ResourceData{}, nil)
-				So(diagnosis, ShouldNotBeEmpty)
-				So(diagnosis[0].Summary, ShouldEqual, expectedError)
-			})
-		})
-	})
-	Convey("Given a resource factory configured with a create timeout", t, func() {
-		createTimeout := 1 * time.Second
-		r := resourceFactory{
-			openAPIResource: &specStubResource{
-				name: "cdn_v1",
-			},
-		}
-		Convey("When crudWithContext is called with the create function (configured to timeout on purpose) and the corresponding timeout identifier", func() {
-			stubCreateFunction := func(data *schema.ResourceData, i interface{}) error {
-				time.Sleep(2 * time.Second)
-				return nil
-			}
-			contextAwareFunc := r.crudWithContext(stubCreateFunction, schema.TimeoutCreate)
-			Convey("Then the returned function which is context aware should time out since the create operation takes longer than the context timeout", func() {
-				ctx := context.Background()
-				ctx, cancel := context.WithTimeout(ctx, createTimeout)
-				defer cancel()
-				diagnosis := contextAwareFunc(ctx, &schema.ResourceData{}, nil)
-				So(diagnosis, ShouldNotBeEmpty)
-				So(diagnosis[0].Summary, ShouldEqual, "context deadline exceeded: 'cdn_v1' create timeout is 20m0s") // the 20m0s is the default timeout if the openAPIResource is not configured with specific timeouts
-			})
-		})
-	})
-}
-
 func TestCreateTerraformResourceSchema(t *testing.T) {
 	Convey("Given a resource factory", t, func() {
 		r, _ := testCreateResourceFactory(t, idProperty, stringProperty)
